More tests written, some corrections made and some helper methods written
Changes:

* Changed ram types in JobConfig to uint64
* Added castToOptional helper methods to Utilities
* Removed old environment conversion from JobGateway
* Corrected some typos in UserGatewayTest
* Wrote more tests for JobGatewayTest and corrected some of the issues with the old tests

+++ src/configfiles/JobConfig.cpp
@@ -168,11 +168,11 @@ JobConfig::~JobConfig() {
 
 }
 
-void JobConfig::set_min_ram(const std::optional <uint32_t> &miB) {
+void JobConfig::set_min_ram(const std::optional <uint64_t> &miB) {
     min_ram_ = miB;
 }
 
-void JobConfig::set_max_ram(const std::optional <uint32_t> &miB) {
+void JobConfig::set_max_ram(const std::optional <uint64_t> &miB) {
     max_ram_ = miB;
 }
 
@@ -221,11 +221,11 @@ void JobConfig::set_backup_ID(std::optional<uint32_t> backupID) {
 }
 
 
-std::optional <uint32_t> &JobConfig::min_ram() {
+std::optional <uint64_t> &JobConfig::min_ram() {
     return min_ram_;
 }
 
-std::optional <uint32_t> &JobConfig::max_ram() {
+std::optional <uint64_t> &JobConfig::max_ram() {
     return max_ram_;
 }
